organized react app into dir and added css modules
