+++ react-app/src/Modal/Modal.module.css
@@ -1,0 +1,3 @@
+.modal {
+    display: flex;
+}
