+++ react-app/src/Modal/index.js
@@ -1,0 +1,3 @@
+import { ModalProvider, Modal } from "./Modal";
+
+export { ModalProvider, Modal };
