+++ react-app/src/components/Login/LoginForm.js
@@ -1,0 +1,66 @@
+import React, { useState } from "react";
+import { useSelector, useDispatch } from "react-redux";
+import { Redirect } from "react-router-dom";
+import { login } from "../../store/session";
+
+import styles from "./LoginForm.module.css";
+// className={styles. }
+
+export default function LoginForm() {
+    const [errors, setErrors] = useState([]);
+    const [email, setEmail] = useState("");
+    const [password, setPassword] = useState("");
+    const user = useSelector((state) => state.session.user);
+    const dispatch = useDispatch();
+
+    const onLogin = async (e) => {
+        e.preventDefault();
+        const data = await dispatch(login(email, password));
+        if (data) {
+            setErrors(data);
+        }
+    };
+
+    const updateEmail = (e) => {
+        setEmail(e.target.value);
+    };
+
+    const updatePassword = (e) => {
+        setPassword(e.target.value);
+    };
+
+    if (user) {
+        return <Redirect to="/" />;
+    }
+
+    return (
+        <form className={styles.form} onSubmit={onLogin}>
+            <div>
+                {errors.map((error, ind) => (
+                    <div key={ind}>{error}</div>
+                ))}
+            </div>
+            <div>
+                <label htmlFor="email">Email</label>
+                <input
+                    name="email"
+                    type="text"
+                    placeholder="Email"
+                    value={email}
+                    onChange={updateEmail}
+                />
+            </div>
+            <div>
+                <label htmlFor="password">Password</label>
+                <input
+                    name="password"
+                    type="password"
+                    placeholder="Password"
+                    value={password}
+                    onChange={updatePassword}
+                />
+                <button type="submit">Login</button>
+            </div>
+        </form>
+    );
+}
