+++ react-app/src/components/Login/LoginForm.module.css
@@ -1,0 +1,3 @@
+.form {
+    display: flex;
+}
