+++ react-app/src/components/Logout/LogoutButton.module.css
@@ -1,0 +1,3 @@
+.button {
+    display: flex;
+}
