+++ react-app/src/components/Logout/index.js
@@ -1,0 +1,3 @@
+import LogoutButton from "./LogoutButton";
+
+export default LogoutButton;
