+++ react-app/src/components/Nav/NavBar.module.css
@@ -1,0 +1,3 @@
+.nav {
+    display: flex;
+}
