+++ react-app/src/components/Nav/index.js
@@ -1,0 +1,3 @@
+import NavBar from "./NavBar";
+
+export default NavBar;
