+++ react-app/src/components/Signup/SignupForm.module.css
@@ -1,0 +1,3 @@
+.form {
+    display: flex;
+}
