+++ react-app/src/components/Signup/index.js
@@ -1,0 +1,3 @@
+import SignupForm from "./SignUpForm";
+
+export default SignupForm;
