+++ react-app/src/components/UserProfile/UserProfile.js
@@ -1,0 +1,39 @@
+import React, { useState, useEffect } from "react";
+import { useParams } from "react-router-dom";
+
+import styles from "./UserProfile.module.css";
+// className={styles. }
+
+export default function UserProfile() {
+    const [user, setUser] = useState({});
+    const { userId } = useParams();
+
+    useEffect(() => {
+        if (!userId) {
+            return;
+        }
+        (async () => {
+            const response = await fetch(`/api/users/${userId}`);
+            const user = await response.json();
+            setUser(user);
+        })();
+    }, [userId]);
+
+    if (!user) {
+        return null;
+    }
+
+    return (
+        <ul className={styles.user}>
+            <li>
+                <strong>User Id</strong> {userId}
+            </li>
+            <li>
+                <strong>Username</strong> {user.username}
+            </li>
+            <li>
+                <strong>Email</strong> {user.email}
+            </li>
+        </ul>
+    );
+}
