+++ react-app/src/components/UserProfile/UserProfile.module.css
@@ -1,0 +1,3 @@
+.user {
+    display: flex;
+}
