+++ react-app/src/components/UserProfile/index.js
@@ -1,0 +1,3 @@
+import UserProfile from "./UserProfile";
+
+export default UserProfile;
