+++ react-app/src/components/UsersList/UsersList.module.css
@@ -1,0 +1,3 @@
+.listContainer {
+    display: flex;
+}
