+++ react-app/src/components/UsersList/index.js
@@ -1,0 +1,3 @@
+import UsersList from "./UsersList";
+
+export default UsersList;
