+++ react-app/src/store/index.js
@@ -2,12 +2,10 @@ import { createStore, combineReducers, applyMiddleware, compose } from "redux";
 import thunk from "redux-thunk";
 import session from "./session";
 import users from "./users";
-import user from "./user";
 
 const rootReducer = combineReducers({
     session,
     users,
-    user,
 });
 
 let enhancer;
